Module Object.Mem

Member maps.

Usually it's better to use Jsont.Object.mem or Jsont.Object.opt_mem directly. But this may be useful in certain abstraction contexts.

type ('o, 'dec) object_map := ('o, 'dec) map

type ('o, 'a) map

The type for mapping a member object to a value 'a stored in an OCaml value of type 'o.

val map : ?doc:string -> ?dec_absent:'a -> ?enc:('o -> 'a) -> ?enc_omit:('a -> bool) -> string -> 'a t -> ('o, 'a) map

See Jsont.Object.mem.

val app : ('o, 'a -> 'b) object_map -> ('o, 'a) map -> ('o, 'b) object_map

app map mmap applies the member map mmap to the contructor of the object map map. In turn this adds the mmap member definition to the object described by map.